aiModels/requirements.txt
# Core FastAPI Framework
fastapi==0.95.2           # High-performance web framework for building APIs with Python.
uvicorn==0.22.0           # ASGI server required to run FastAPI apps in development and production environments.

# Pydantic for Data Validation
pydantic==1.10.7          # Data validation and management library used natively in FastAPI for request/response models.

# Authentication and Security
python-jose==3.3.0        # Library for generating and validating JWT tokens, essential for user authentication.
bcrypt==4.0.1             # Secure hashing library for encrypting user passwords before storing in the database.

# Database Libraries
sqlalchemy==2.0.20        # Object-Relational Mapper (ORM) for managing and querying relational databases in Python.
databases==0.6.1          # Async database support for working seamlessly with SQLAlchemy in FastAPI projects.
psycopg2-binary==2.9.6    # PostgreSQL adapter, required if you're using PostgreSQL as the database backend.
sqlite-utils==3.32.1      # Simplifies interaction with SQLite for local development and testing purposes.

# Redis for Caching and Real-Time Features
redis==5.1.0              # Redis client for managing caching and real-time data like notifications or session storage.
aioredis==2.0.1           # Async support for Redis, enabling integration with FastAPI's async architecture.

# Logging and Monitoring
loguru==0.7.0             # Advanced logging library for clean, structured logs with enhanced debugging features.
sentry-sdk==1.28.1        # SDK for capturing application errors and monitoring performance with Sentry.

# Machine Learning and AI
scikit-learn==1.2.2       # Core machine learning library for implementing models like SLA predictions and engineer matching.
numpy==1.25.0             # High-performance numerical computation library, essential for ML model training.
pandas==2.1.1             # Data manipulation library used for processing structured datasets (e.g., task and engineer data).
joblib==1.3.2             # Efficient saving/loading of trained machine learning models, used to persist AI models in production.
xgboost==1.7.6            # Gradient boosting library for high-performance machine learning models, particularly for SLA predictions.

# WebSocket and Real-Time Communication
websockets==11.0.3        # Library for WebSocket support, used for real-time communication with engineers.

# Email Notifications
fastapi-mail==1.2.1       # Provides easy email integration with FastAPI for features like task updates or SLA alerts.
email-validator==1.3.0    # Validates email addresses during registration or when sending notifications to users.

# Task Scheduling
apscheduler==3.10.1       # Advanced scheduler for managing periodic jobs like SLA monitoring, notifications, and model retraining.

# Testing and Development
pytest==7.4.2             # Testing framework for writing unit tests, integration tests, and functional tests.
pytest-asyncio==0.21.0    # Adds asyncio support to pytest, enabling tests for async FastAPI endpoints.
httpx==0.24.1             # Async HTTP client for making API calls during testing or internal service communication.

# Environment and Configuration Management
python-dotenv==1.0.0      # Loads environment variables from a `.env` file, simplifying configuration for local and production environments.

# Data Visualisation (Optional for Dashboards)
matplotlib==3.7.3         # Plotting library used for creating visual dashboards (e.g., SLA compliance trends).
seaborn==0.12.2           # Statistical data visualisation library, often used for exploring trends in system data.

# Optional for Deployment
gunicorn==20.1.0          # WSGI HTTP server, used for deploying FastAPI applications in production environments.
PyJWT==2.8.0
